#### Check List:
- [x ] Tests pass (`./test.sh ebarlas` shows no differences between expected and actual outputs)
- [ x] All formatting changes by the build are committed
- [ x] Your launch script is named `calculate_average_ebarlas.sh` (make sure to match casing of your GH user name) and is executable
- [x ] Output matches that of `calculate_average_baseline.sh`
* Execution time: 5.9s on `c5.2xlarge` down from 6.8s (13% reduction)
* Execution time of reference implementation: 3m59s

Execution time on CCX33 VM down from 5.7s to 4.7s (18% reduction)