@ebarlas 2024-01-10
Hm, that's odd. It works for me. Are you sure there's an issue? I'll look again.

```
$ ./test.sh ebarlas
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-1.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-10.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-10000-unique-keys.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-2.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-20.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-3.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-boundaries.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-complex-utf8.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-dot.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-short.txt
Validating calculate_average_ebarlas.sh -- src/test/resources/samples/measurements-shortest.txt
```