@gunnarmorling 2024-01-10
Thx, looking good now.