@gunnarmorling 2024-01-10
Nice, this takes it from 00:05.683 down to 00:04.996 on the new evaluation environment.